Revise User save, Add user/userIdx api

+++ dosports/src/main/java/umc/dosports/User/UserForm.java
@@ -67,7 +67,7 @@ public class UserForm {
         this.gender = gender;
     }
 
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
